backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pandas>=1.3.3
scikit-learn>=0.24.2
nltk>=3.6.3
spacy>=3.1.3
pydantic>=1.8.2
python-jose>=3.3.0
passlib>=1.7.4